+++ src/renderer/document/CanvasToolSelectText.ts
@@ -4,8 +4,8 @@ export class CanvasToolSelectText extends CanvasTool {
 
   public override idleCursor = 'default';
 
-  public override canSetStrokeWidth = false;
-  public override canSetColor = false;
+  public override canSetStrokeWidth = true;
+  public override canSetColor = true;
 
   override onDeselect() {
     document.getSelection().removeAllRanges();